requirements.txt
flask==3.0.0
werkzeug==3.0.0
faster-whisper==1.2.1
librosa==0.11.0
numpy<2
scipy==1.11.0
soundfile==0.12.1
torch==2.1.2
torchaudio==2.1.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0